refactor(oxlint-rules): migrate custom rules to TypeScript with @oxlint/plugins
Switches the four custom oxlint rules from `.mjs` with eslint-typed JSDoc
to `.ts` with first-party `@oxlint/plugins` types, matching the official
oxlint plugin authoring style and removing the eslint type dependency.
