+++ package.json
@@ -50,6 +50,7 @@
     "@babel/core": "7.29.0",
     "@faker-js/faker": "10.4.0",
     "@next/env": "16.2.4",
+    "@oxlint/plugins": "1.62.0",
     "@rolldown/plugin-babel": "0.2.3",
     "@storybook/addon-a11y": "10.3.5",
     "@storybook/addon-docs": "10.3.5",
