+++ tsconfig.json
@@ -30,7 +30,6 @@
     "next.config.ts",
     "oxlint.config.ts",
     "postcss.config.mjs",
-    "scripts/**/*.mjs",
     "src/**/*.ts",
     "src/**/*.tsx",
     "vitest.config.ts",
